+++ centit-utils/src/main/java/com/centit/support/common/ICachedObject.java
@@ -2,7 +2,7 @@ package com.centit.support.common;
 
 public interface ICachedObject<T> {
     //单位分钟，表示一个月
-    int NOT_REFRESH_PERIOD = 43200;
+    long NOT_REFRESH_PERIOD = 43200L;
 
     void evictCahce();
